Add general Ctrl+V paste that works very well.
I could not make deselecting nodes work well by clicking on empty space (seems mouse filter Stop is broken) so I made the RMB de-select.

Also Ctrl+V when a node is selected will overwrite the image or link inside, will add a popup confirmation later. Does not overwrite body text, need to add some behaviour for that perhaps (if text edit is not in focus, replace if it's empty).