Add script to automatically expand RichTextLabels to right size
Unfortunately I can't use this to automatically update the size of the TextEdit node since the RTL is hidden so it only resizes when shown.
So the existing script for ExpandingTextEdit is still needed even if it's buggy.

Also the new script to expand the RTL always expands it vertically to avoid having a scroll bar and doesn't have any limits which will need to be added.

Ideally RTL would have a defined limit and the TextEdit would not have any limits so the entire text can be seen while editing.

One option is to have the TextEdit be outside the node and just popup when needed to write, this way it grow without affecting the size of the original box. Can possible have it popup slightly offset as well so users can the resulting bbcode directly in the RTL